# Add CLI argument for specifying CiliumNetworkPolicies in Performance Tests